Refactor BulkUploadTrack component to modularize functionality and improve user experience
- Removed the original BulkUploadTrack implementation and replaced it with a modular structure, separating concerns into distinct components for DropZone, TrackList, UnmatchedArt, and UploadProgress.
- Introduced hooks for file processing and drag-and-drop functionality, enhancing code organization and maintainability.
- Improved state management for handling file uploads, cover art matching, and user interactions, ensuring a smoother experience during bulk uploads.
- Updated UI components to provide clearer feedback on upload progress and track management, enhancing overall usability.
