+++ packages/frontend/src/pages/BulkUploadTrack/hooks/useDragAndDrop.ts
@@ -1,0 +1,53 @@
+import { useCallback, useRef, useState } from "react";
+import { DragEvent } from "react";
+
+export function useDragAndDrop(onDrop: (files: File[]) => void) {
+  const [isDragging, setIsDragging] = useState(false);
+  const dragCounter = useRef(0);
+
+  const handleDragIn = useCallback((e: DragEvent<HTMLDivElement>) => {
+    e.preventDefault();
+    e.stopPropagation();
+    dragCounter.current++;
+    if (e.dataTransfer?.items && e.dataTransfer.items.length > 0) {
+      setIsDragging(true);
+    }
+  }, []);
+
+  const handleDragOut = useCallback((e: DragEvent<HTMLDivElement>) => {
+    e.preventDefault();
+    e.stopPropagation();
+    dragCounter.current--;
+    if (dragCounter.current === 0) {
+      setIsDragging(false);
+    }
+  }, []);
+
+  const handleDragOver = useCallback((e: DragEvent<HTMLDivElement>) => {
+    e.preventDefault();
+    e.stopPropagation();
+  }, []);
+
+  const handleDrop = useCallback(
+    (e: DragEvent<HTMLDivElement>) => {
+      e.preventDefault();
+      e.stopPropagation();
+      setIsDragging(false);
+      dragCounter.current = 0;
+
+      const files = Array.from(e.dataTransfer?.files || []);
+      onDrop(files);
+    },
+    [onDrop]
+  );
+
+  return {
+    isDragging,
+    handlers: {
+      handleDragIn,
+      handleDragOut,
+      handleDragOver,
+      handleDrop,
+    },
+  };
+}
